taskList

+++ src/main/java/Tproject/service/TaskServiceImpl.java
@@ -2,7 +2,9 @@ package Tproject.service;
 
 import Tproject.dto.TaskCreateDto;
 import Tproject.model.Task;
+import Tproject.model.TaskList;
 import Tproject.model.User;
+import Tproject.repository.TaskListRepository;
 import Tproject.repository.TaskRepository;
 import Tproject.repository.UserRepository;
 import Tproject.util.JwtUtil;
@@ -22,14 +24,19 @@ import java.util.List;
 public class TaskServiceImpl implements TaskService{
     private final TaskRepository taskRepository;
     private final UserRepository userRepository;
+    private final TaskListRepository taskListRepository;
     private final JwtUtil jwtUtil;
     private final UserUtil userUtil;
     @Override
-    public Task create(TaskCreateDto createDto,HttpServletRequest request) {
+    public Task create(Long taskListId,TaskCreateDto createDto,HttpServletRequest request) {
+        TaskList taskList = taskListRepository.findById(taskListId)
+                .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND,"Список не найден"));
+        if (!taskList.getUser().getId().equals(userUtil.getUserByRequest(request).getId())) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN);
+        }
+
         Task newTask = new Task();
-        newTask.setCreator(userUtil.getUserByRequest(request));
-        newTask.setExecutor(userRepository.findByUsername(createDto.getExecutor().getUsername())
-                .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND,"Пользователь не найден")));
+        newTask.setList(taskList);
         newTask.setDeadline(createDto.getDeadline());
         newTask.setDescription(createDto.getDescription());
         newTask.setTitle(createDto.getTitle());
@@ -39,28 +46,36 @@ public class TaskServiceImpl implements TaskService{
 
 
     @Override
-    public List<Task> getTasksByRequest(HttpServletRequest request) {
-        User user = userUtil.getUserByRequest(request);
-        return taskRepository.findByExecutor(user);
+    public List<Task> getByTaskListId(Long taskListId) {
+        TaskList taskList = taskListRepository.findById(taskListId)
+                .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND));
+
+        return taskList.getTasks();
     }
 
     @Override
-    public Task update(Long id, TaskCreateDto createDto) {
+    public Task update(Long id, TaskCreateDto createDto,HttpServletRequest request) {
         Task task = taskRepository.findById(id)
                 .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND,"Задача не найдена"));
+        if (!task.getList().getUser().getId().equals(userUtil.getUserByRequest(request).getId())) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN);
+        }
+
         task.setTitle(createDto.getTitle());
         task.setDescription(createDto.getDescription());
         task.setDeadline(createDto.getDeadline());
-        task.setExecutor(userRepository.findByUsername(createDto.getExecutor().getUsername())
-                .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND,"Пользователь не найден")));
         taskRepository.save(task);
         return task;
     }
 
     @Override
-    public Task changeStatus(Long id, boolean isDone) {
+    public Task changeStatus(Long id, boolean isDone,HttpServletRequest request) {
         Task task = taskRepository.findById(id)
                 .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND,"Задача не найдена"));
+        if (!task.getList().getUser().getId().equals(userUtil.getUserByRequest(request).getId())) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN);
+        }
+
         if(task.isDone() == isDone){
             return task;
         }
@@ -70,9 +85,13 @@ public class TaskServiceImpl implements TaskService{
     }
 
     @Override
-    public String deleteTask(Long id) {
+    public String deleteTask(Long id,HttpServletRequest request) {
         Task task = taskRepository.findById(id)
                 .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND,"Задача не найдена"));
+        if (!task.getList().getUser().getId().equals(userUtil.getUserByRequest(request).getId())) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN);
+        }
+
         taskRepository.delete(task);
         return "Удалено";
     }
